add EmailManager to deport logic build & send
