+++ masterpiece-server/src/main/java/fr/formation/masterpiece/api/services/SharingSessionService.java
@@ -15,8 +15,5 @@ public interface SharingSessionService {
 
     List<SharingSessionViewDto> getAllSessions();
 
-    void buildSessionMail(SharingSessionViewDto dtoToReturn)
-            throws MessagingException;
-
     boolean isDateValid(LocalDateTime date);
 }
