+++ masterpiece-server/src/main/java/fr/formation/masterpiece/api/services/impl/SharingSessionServiceImpl.java
@@ -1,8 +1,6 @@
 package fr.formation.masterpiece.api.services.impl;
 
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.mail.MessagingException;
@@ -13,14 +11,12 @@ import org.springframework.stereotype.Service;
 import fr.formation.masterpiece.api.repositories.CustomUserRepository;
 import fr.formation.masterpiece.api.repositories.SharingSessionRepository;
 import fr.formation.masterpiece.api.repositories.SubjectRepository;
-import fr.formation.masterpiece.api.services.EmailService;
 import fr.formation.masterpiece.api.services.SharingSessionService;
 import fr.formation.masterpiece.commons.config.AbstractService;
 import fr.formation.masterpiece.commons.exceptions.ResourceNotFoundException;
 import fr.formation.masterpiece.domain.dtos.SharingSessionCreateDto;
 import fr.formation.masterpiece.domain.dtos.views.SharingSessionViewDto;
 import fr.formation.masterpiece.domain.entities.CustomUser;
-import fr.formation.masterpiece.domain.entities.Mail;
 import fr.formation.masterpiece.domain.entities.SharingSession;
 import fr.formation.masterpiece.domain.entities.Subject;
 import fr.formation.masterpiece.security.SecurityHelper;
@@ -35,16 +31,12 @@ public class SharingSessionServiceImpl extends AbstractService
 
     private final SharingSessionRepository sharingSessionRepository;
 
-    private final EmailService emailService;
-
     public SharingSessionServiceImpl(CustomUserRepository userRepository,
             SubjectRepository subjectRepository,
-            SharingSessionRepository sharingSessionRepository,
-            EmailService emailService) {
+            SharingSessionRepository sharingSessionRepository) {
 	this.userRepository = userRepository;
 	this.subjectRepository = subjectRepository;
 	this.sharingSessionRepository = sharingSessionRepository;
-	this.emailService = emailService;
     }
 
     @Override
@@ -65,25 +57,6 @@ public class SharingSessionServiceImpl extends AbstractService
     }
 
     @Override
-    public void buildSessionMail(SharingSessionViewDto session)
-            throws MessagingException {
-	List<String> recipients = getRecipients();
-	StringBuilder builder = new StringBuilder();
-	String content = builder.append("New session booked").append("<p>")
-	        .append("Date : " + formatDate(session.getStartTime()))
-	        .append("<p>")
-	        .append("Starting time : " + formatTime(session.getStartTime()))
-	        .append("<p>")
-	        .append("Ending time : " + formatTime(session.getEndTime()))
-	        .append("<p>").append(" The lecturer will be : "
-	                + session.getUser().getUsername())
-	        .toString();
-	Mail mail = new Mail("SyK", recipients,
-	        "New session : " + session.getSubject().getTitle(), content);
-	emailService.sendMail(mail);
-    }
-
-    @Override
     public List<SharingSessionViewDto> getAllSessions() {
 	List<SharingSessionViewDto> list = convertList(
 	        sharingSessionRepository.getAllSessionWithUserEnable(),
@@ -96,23 +69,6 @@ public class SharingSessionServiceImpl extends AbstractService
 	return !sharingSessionRepository.existsByStartTimeBetween(
 	        setToTime(dateTime, 00, 00, 00),
 	        setToTime(dateTime, 23, 59, 59));
-    }
-
-    private List<String> getRecipients() {
-	List<CustomUser> users = userRepository.findAll();
-	List<String> recipients = new ArrayList<>();
-	users.forEach(user -> recipients.add(user.getEmail()));
-	return recipients;
-    }
-
-    private String formatDate(LocalDateTime dateTime) {
-	DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
-	return dateTime.format(formatter);
-    }
-
-    private String formatTime(LocalDateTime dateTime) {
-	DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH-mm");
-	return dateTime.format(formatter);
     }
 
     /**
